requirements.txt
anthropic>=0.25.0
openai>=1.30.0
elevenlabs>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0
pydub>=0.25.1